Implements #2 - Define lock order
When locks are created via g_lock_create_... each lock is given an
index. When a session attempts to take a lock it is validated that
the lock being taken is in the proper order. If it is not then the
program will abort with the thought that it is better for the program
to exit during development then in production.

Whether the program is aborted can be set via G_LOCK_ORDER_ABORT.

A lock is allowed to be taken if:
- It has not already taken the same indexed lock.
  - This does not apply to recursive locks
- The lock index to be taken is not less than the locks already present
  in the session. This way the order any locks are taken are always the
  same irrespective of a given thread.